Chore : 의존성 주입 방법 변경 및 자잘한 수정
의존성 주입 시 private final 키워드와 @RequiredArgsConstructor 사용
변수명, import 추가, 간단한 주석 추가 등

+++ src/main/java/inbox/inbox/portfolio/PortfolioEmailConfirmRepository.java
@@ -3,7 +3,9 @@ package inbox.inbox.portfolio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+// DB에 접근하게 해줄 인터페이스
 @Repository
-public interface PortfolioEmailConfirmRepository extends JpaRepository<PortfolioEmailConfirm,Integer> {
+public interface PortfolioEmailConfirmRepository extends
+    JpaRepository<PortfolioEmailConfirm, Long> {
 
 }
